# Introduce StreamingWithSIMDOptimized mode for PQ distance calculator