--- a/rs/quantization/src/lib.rs
+++ b/rs/quantization/src/lib.rs
@@ -1,3 +1,4 @@
+#![feature(portable_simd)]
 pub mod pq;
 pub mod pq_builder;
 pub mod quantization;